requirements.txt
boto3
Pillow
opencv-python
pywin32